refactor to move to existing suite
