+++ pkg/test/framework/components/prometheus/prometheus.go
@@ -16,7 +16,6 @@ package prometheus
 
 import (
 	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
-	"github.com/prometheus/common/model"
 	prom "github.com/prometheus/common/model"
 
 	"istio.io/istio/pkg/test"
@@ -47,7 +46,7 @@ type Instance interface {
 	Sum(val prom.Value, labels map[string]string) (float64, error)
 	SumOrFail(t test.Failer, val prom.Value, labels map[string]string) float64
 
-	KnownMetrics() (model.LabelValues, error)
+	KnownMetrics() (prom.LabelValues, error)
 }
 
 type Config struct {
