+++ tools/packaging/common/envoy_bootstrap.json
@@ -118,20 +118,28 @@
       },
       {
         "tag_name": "source_zone",
-        "regex": "^zone\\.((.+?)\\.)",
+        "regex": "zone\\.((.+?)\\.)",
       },
       {
         "tag_name": "destination_zone",
-        "regex": "^zone\\.((.+?)\\.)",
+        "regex": "zone\\.((.+?)\\.)",
       },
       {
         "tag_name": "worker",
-        "regex": "^worker\\.((.+?)\\.)",
+        "regex": "worker\\.((.+?)\\.)",
       },
       {
         "tag_name": "http_prefix",
         "regex": "^http\\.((.*?)\\.)\\w+?$"
-      }
+      },
+      {
+        "tag_name": "curve",
+        "regex": "curves\\.((.+?)\\.)"
+      },
+      {
+        "tag_name": "cipher",
+        "regex": "ciphers\\.((.+?)\\.)"
+      },
     ],
     "stats_matcher": {
       "inclusion_list": {
